Report
In this evaluation on ABAQUS, our aim is to model a prosthetic implant
for a hip replacement. Indeed, prosthetic implants needed to be firmly
attached to the rest of the skeleton and thus we must look, with a
model, at the deformations that can appear in the implant. In order to
do this, we must first create the 3D elements and then enter the
characteristics of the elements in ABAQUS.